migration to Xtext 2.2.0

+++ de.bmw.carit.jnario.spec.tests/xtend-gen/calculator/Calculator.java
@@ -5,7 +5,7 @@ import org.eclipse.xtext.xbase.lib.IntegerExtensions;
 @SuppressWarnings("all")
 public class Calculator {
   public int add(final int a, final int b) {
-    int _operator_plus = IntegerExtensions.operator_plus(((Integer)a), ((Integer)b));
+    int _operator_plus = IntegerExtensions.operator_plus(a, b);
     return _operator_plus;
   }
 }
